Remove LazyServiceEntityRepository
It was only useful for ORM 2. This allows to inline
ServiceEntityRepositoryProxy inside ServiceEntityRepository.
