+++ phpcs.xml.dist
@@ -28,7 +28,6 @@
     <rule ref="PSR1.Classes.ClassDeclaration.MultipleClasses">
         <exclude-pattern>tests/*</exclude-pattern>
         <exclude-pattern>src/Repository/RepositoryFactoryCompatibility.php</exclude-pattern>
-        <exclude-pattern>src/Repository/ServiceEntityRepository.php</exclude-pattern>
     </rule>
     <rule ref="Squiz.Classes.ClassFileName.NoMatch">
         <exclude-pattern>tests/*</exclude-pattern>
